Extracted autoscroll into hook
