+++ stories/DatePicker.stories.js
@@ -1,5 +1,4 @@
-import React, { useState } from 'react';
-import { action } from '@storybook/addon-actions';
+import React from 'react';
 import Harness from './helpers/Harness';
 import DatePicker from '../src/components/Fields/DatePicker';
 import '../src/styles/_all.scss';
@@ -13,23 +12,19 @@ const requiredProps = {
 export default { title: 'DatePicker' };
 
 export const Field = () => (
-  <Harness
-    requiredProps={requiredProps}
-  >
+  <Harness requiredProps={requiredProps}>
     {props => <DatePicker {...props} />}
   </Harness>
 );
 
 export const AutoScroll = () => (
-  <Harness
-    requiredProps={requiredProps}
-  >
+  <Harness requiredProps={requiredProps}>
     {props => (
-      <div style={{ backgroundColor: 'silver', height: '400px', overflowY: 'scroll' }}>
-        <div style={{ padding: '330px 0' }}>
+      <div style={{ backgroundColor: '#e7e7e7', height: '400px', overflowY: 'scroll' }}>
+        <div style={{ padding: '300px 0' }}>
           <DatePicker {...props} />
         </div>
       </div>
     )}
   </Harness>
-  );
+);
